+++ gridgo-bean/pom.xml
@@ -1,4 +1,6 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 	<artifactId>gridgo-bean</artifactId>
 
@@ -13,6 +15,7 @@
 		<msgpack.version>0.8.16</msgpack.version>
 		<jsonsmart.version>2.3</jsonsmart.version>
 		<evalex.version>2.0</evalex.version>
+		<junit.version>4.12</junit.version>
 	</properties>
 
 	<dependencies>
@@ -38,6 +41,13 @@
 			<groupId>com.udojava</groupId>
 			<artifactId>EvalEx</artifactId>
 			<version>${evalex.version}</version>
+		</dependency>
+
+		<dependency>
+			<groupId>junit</groupId>
+			<artifactId>junit</artifactId>
+			<version>${junit.version}</version>
+			<scope>test</scope>
 		</dependency>
 	</dependencies>
 </project>
